Added heuristics for Astar. Added Astar animator, added diagonals for a star added frame/state animation

+++ src/pathanimations/astar/heuristics/Heuristic.java
@@ -3,5 +3,5 @@ package pathanimations.astar.heuristics;
 import graph.Point;
 
 public interface Heuristic {
-    int getFor(Point p1, Point p2);
+    float getFor(Point p1, Point p2);
 }
